Replace broken symlinks with links.
If someone asks to link a binary, and that link exists, but
is broken, replace it, rather than blowing up.

Maybe this behavior will be unnecessary once links are
tracked 'globally' and unlinked when environments are
removed (#69), but until then, this case happens if say,
a link is moved from one env to another.