Start dealing in 'effective' configurations.
Gets us closer towards converging venvs in the
right cases even when e.g. a package moves from
explicit installation into a bundle.

And when I say 'closer', what I mean is -- this
patch temporarily breaks everything entirely,
since we need to change to format we serialize a
converged venv in [installed.toml], since as is
after this commit, we always think a venv is not
the same as what's on disk.

Cache invalidation, the most difficult problem in
programming.

OOPs.