And now fix thinking nothing is ever converged.
Compare the post-modified config (via JSON) with the existing
one.

This is because we still modify the config in the process of
writing it out (by adding a sys.version key).

The real fix is to actually stop doing that...

But for now since master is currently 'broken' vis a vis
caching, just put in the quickfix and be ready for it to
bite in the ass later.